Add files via upload
Project Overview

Definition:
Little Readers is an online children’s library careering to the literary needs of young readers. We offer a wide variety of books ranging from alphabet books for toddlers to works of world-famous authors like Roald Dahl, Enid Blyton and J.K Rowling. And, best of all, we deliver them right to your doorstep!

Main Menu:
The Main Menu offers options to become a member and to login to your pre-existing account.
•	In order to create a membership, the child must not be above 13 years of age. The parent/guardian must fill in details such as their child’s name, date of birth, parent/guardian name, email id, address etc. The parent/guardian must then create their account password and pay for a membership of their choice.
•	In order to login to a pre-existing account, the user must enter their correct username and password.
Home:
Once the login is successful, the program moves on to Home, where there are two options: Browse Books and Logout.
•	In Browse Books, the user will be able to view our collections. Upon selecting a collection, he/she will be able to view the books in that collection. A user may only borrow one book at a time.
•	In Logout, the user will be ablt to log out of his/her Little Reader account.
Collections:
In Collections, the user will be able to view the different sections into which our books are divided. This division helps us arrange our books in an orderly manner.
Books:
Each section contains books belonging to the same. A user may select one book to borrow. Upon selection, the user may either exchange the book for another, or proceed to Logout.
Delivery:
Books will be delivered within a span of 36 hours.